api/requirements.txt
pydantic
langchain
langchain-google-genai
python-dotenv
google-generativeai
fastapi
uvicorn
